ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
xgboost==2.0.3
prophet==1.1.5
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23

